v1.0.3 Code review cleanup v1.0.2 Persist history options

+++ battery.h
@@ -34,8 +34,8 @@ public:
 private:
     struct PSoC
     {
-        double psoc;
-        double v;
+        float psoc;
+        float v;
     };
 
     float readVoltage();
