+++ debounced_button.h
@@ -25,7 +25,7 @@ private:
     int m_pin = 0;
     int m_activeState = HIGH;
 
-    uint16_t m_lastChange = 0;
+    uint32_t m_lastChange = 0;
     int m_lastValue = LOW;
     int m_buttonValue = LOW;
 };
